Android: fix the pointer-through-int sites the suite does not reach
The Phase 3 harvest (279 explicit pointer->u32 casts) left a handful that
truncate the address of a global or pass a pointer through an int
parameter, on paths the golden suite never runs. Under PIE each is a
crash waiting for a device:

- mnsnap.c: `(u32) &mnSnap_804A0A10 + off` (12 sites) and `(u32) snap`
  arithmetic -> uintptr_t (the snapshot menu)
- grmutecity.c: `(u32) grMc_8049F440 + off` -> uintptr_t (Mute City load)
- ifmagnify.c: `(u32) &ifMagnify_804A1DE0` -> uintptr_t
- gm_16AE.c / gmregclear.c: the 1-P match-end record travelled through
  `int` into fn_8016C46C -> fn_8017E8A4/fn_8017EE40; intptr_t end to end
- particle.c: hsd_80394544's font table parameter was s32; intptr_t

uintptr_t/intptr_t are the 32-bit types on GCN, so the decomp's codegen
is unchanged. tools/pc_audio_check.py honours MELEE_SUITE_PORT like the
suite does; the PIE binary's music is bit-exact (corr 1.0000, constant
lag), and both suites are within baseline.

+++ src/melee/gm/gmregclear.c
@@ -1616,7 +1616,7 @@ typedef struct {
 } u8_bits;
 
 #pragma dont_inline on
-void fn_8017E8A4(int arg0_int)
+void fn_8017E8A4(intptr_t arg0_int)
 {
     MatchEnd* arg0 = (MatchEnd*) arg0_int;
     lbl_8046B6A0_t* gm = gm_16AE_GetUnkData_0();
@@ -1786,7 +1786,7 @@ bool fn_8017EDDC(void)
 }
 
 #pragma dont_inline on
-void fn_8017EE40(int arg0_int)
+void fn_8017EE40(intptr_t arg0_int)
 {
     MatchEnd* arg0 = (MatchEnd*) arg0_int;
     struct StartMeleeRules* rules;
